Graphs!
Box plots are over a range to give each country a chance to represent their data (e.g. if had value for 2014 but not 2015) and averaged so no country has a higher weighting over another. The first two are with the data set provided looking at youth NEET %. The next two are on enrolment in education in the same time periods so hopefully see where there are similarities in the changes on IQ range or medians (remember that they SHOULD have negative correlation).
The next two are line graphs for youth NEET and youth education enrolment over the last ~35 years so we can visually assess the differences between trends.